moved deprecated waveform widgets and renderers to folder deprecrated and dont add to the preferences list
