+++ CMakeLists.txt
@@ -1206,12 +1206,12 @@ add_library(mixxx-lib STATIC EXCLUDE_FROM_ALL
   src/util/workerthreadscheduler.cpp
   src/util/xml.cpp
   src/waveform/guitick.cpp
-  src/waveform/renderers/glslwaveformrenderersignal.cpp
-  src/waveform/renderers/glvsynctestrenderer.cpp
-  src/waveform/renderers/glwaveformrenderbackground.cpp
-  src/waveform/renderers/glwaveformrendererfilteredsignal.cpp
-  src/waveform/renderers/glwaveformrendererrgb.cpp
-  src/waveform/renderers/glwaveformrenderersimplesignal.cpp
+  src/waveform/renderers/deprecated/glslwaveformrenderersignal.cpp
+  src/waveform/renderers/deprecated/glvsynctestrenderer.cpp
+  src/waveform/renderers/deprecated/glwaveformrenderbackground.cpp
+  src/waveform/renderers/deprecated/glwaveformrendererfilteredsignal.cpp
+  src/waveform/renderers/deprecated/glwaveformrendererrgb.cpp
+  src/waveform/renderers/deprecated/glwaveformrenderersimplesignal.cpp
   src/waveform/renderers/waveformmark.cpp
   src/waveform/renderers/waveformmarkrange.cpp
   src/waveform/renderers/waveformmarkset.cpp
@@ -1238,12 +1238,12 @@ add_library(mixxx-lib STATIC EXCLUDE_FROM_ALL
   src/waveform/waveformmarklabel.cpp
   src/waveform/waveformwidgetfactory.cpp
   src/waveform/widgets/emptywaveformwidget.cpp
-  src/waveform/widgets/glrgbwaveformwidget.cpp
-  src/waveform/widgets/glsimplewaveformwidget.cpp
-  src/waveform/widgets/glslwaveformwidget.cpp
-  src/waveform/widgets/glvsynctestwidget.cpp
-  src/waveform/widgets/glwaveformwidget.cpp
-  src/waveform/widgets/glwaveformwidgetabstract.cpp
+  src/waveform/widgets/deprecated/glrgbwaveformwidget.cpp
+  src/waveform/widgets/deprecated/glsimplewaveformwidget.cpp
+  src/waveform/widgets/deprecated/glslwaveformwidget.cpp
+  src/waveform/widgets/deprecated/glvsynctestwidget.cpp
+  src/waveform/widgets/deprecated/glwaveformwidget.cpp
+  src/waveform/widgets/deprecated/glwaveformwidgetabstract.cpp
   src/waveform/widgets/hsvwaveformwidget.cpp
   src/waveform/widgets/rgbwaveformwidget.cpp
   src/waveform/widgets/softwarewaveformwidget.cpp
