+++ src/waveform/renderers/deprecated/glslwaveformrenderersignal.cpp
@@ -1,4 +1,4 @@
-#include "waveform/renderers/glslwaveformrenderersignal.h"
+#include "waveform/renderers/deprecated/glslwaveformrenderersignal.h"
 #if !defined(QT_NO_OPENGL) && !defined(QT_OPENGL_ES_2)
 
 #ifdef MIXXX_USE_QOPENGL
