+++ src/waveform/renderers/deprecated/glwaveformrenderer.h
