+++ src/waveform/renderers/deprecated/glwaveformrendererrgb.cpp
@@ -1,9 +1,9 @@
-#include "waveform/renderers/glwaveformrendererrgb.h"
+#include "waveform/renderers/deprecated/glwaveformrendererrgb.h"
 #if !defined(QT_NO_OPENGL) && !defined(QT_OPENGL_ES_2)
 
 #include "util/math.h"
+#include "waveform/renderers/waveformwidgetrenderer.h"
 #include "waveform/waveform.h"
-#include "waveformwidgetrenderer.h"
 
 namespace {
 const float kHeightScaleFactor = 255.0f / sqrtf(255 * 255 * 3);
