+++ src/waveform/renderers/deprecated/glwaveformrendererrgb.h
