+++ src/waveform/renderers/deprecated/glwaveformrenderersimplesignal.h
