+++ src/waveform/waveformwidgetfactory.cpp
@@ -40,12 +40,12 @@
 #include "waveform/widgets/qtvsynctestwidget.h"
 #include "waveform/widgets/qtwaveformwidget.h"
 #endif
+#include "waveform/widgets/deprecated/glrgbwaveformwidget.h"
+#include "waveform/widgets/deprecated/glsimplewaveformwidget.h"
+#include "waveform/widgets/deprecated/glslwaveformwidget.h"
+#include "waveform/widgets/deprecated/glvsynctestwidget.h"
+#include "waveform/widgets/deprecated/glwaveformwidget.h"
 #include "waveform/widgets/emptywaveformwidget.h"
-#include "waveform/widgets/glrgbwaveformwidget.h"
-#include "waveform/widgets/glsimplewaveformwidget.h"
-#include "waveform/widgets/glslwaveformwidget.h"
-#include "waveform/widgets/glvsynctestwidget.h"
-#include "waveform/widgets/glwaveformwidget.h"
 #include "waveform/widgets/hsvwaveformwidget.h"
 #include "waveform/widgets/rgbwaveformwidget.h"
 #include "waveform/widgets/softwarewaveformwidget.h"
@@ -55,6 +55,26 @@
 #include "widget/wwaveformviewer.h"
 
 namespace {
+bool isDeprecated(WaveformWidgetType::Type t) {
+    switch (t) {
+    case WaveformWidgetType::GLRGBWaveform:
+        return true;
+    case WaveformWidgetType::GLSimpleWaveform:
+        return true;
+    case WaveformWidgetType::GLSLFilteredWaveform:
+        return true;
+    case WaveformWidgetType::GLSLRGBWaveform:
+        return true;
+    case WaveformWidgetType::GLSLRGBStackedWaveform:
+        return true;
+    case WaveformWidgetType::GLVSyncTest:
+        return true;
+    case WaveformWidgetType::GLFilteredWaveform:
+        return true;
+    default:
+        return false;
+    }
+}
 // Returns true if the given waveform should be rendered.
 bool shouldRenderWaveform(WaveformWidgetAbstract* pWaveformWidget) {
     if (pWaveformWidget == nullptr ||
@@ -1063,7 +1083,7 @@ void WaveformWidgetFactory::evaluateWidgets() {
             continue;
         }
 
-        bool active = true;
+        bool active = !isDeprecated(static_cast<WaveformWidgetType::Type>(type));
         if (isOpenGlAvailable()) {
             if (useOpenGles && !useOpenGl) {
                 active = false;
