+++ src/waveform/widgets/deprecated/glwaveformwidget.cpp
@@ -1,11 +1,11 @@
-#include "waveform/widgets/glwaveformwidget.h"
+#include "waveform/widgets/deprecated/glwaveformwidget.h"
 
 #include <QPainter>
 
 #include "moc_glwaveformwidget.cpp"
 #include "util/performancetimer.h"
-#include "waveform/renderers/glwaveformrenderbackground.h"
-#include "waveform/renderers/glwaveformrendererfilteredsignal.h"
+#include "waveform/renderers/deprecated/glwaveformrenderbackground.h"
+#include "waveform/renderers/deprecated/glwaveformrendererfilteredsignal.h"
 #include "waveform/renderers/waveformrenderbeat.h"
 #include "waveform/renderers/waveformrendererendoftrack.h"
 #include "waveform/renderers/waveformrendererpreroll.h"
