+++ src/waveform/widgets/deprecated/glwaveformwidget.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "waveform/widgets/glwaveformwidgetabstract.h"
+#include "waveform/widgets/deprecated/glwaveformwidgetabstract.h"
 
 class GLWaveformWidget : public GLWaveformWidgetAbstract {
     Q_OBJECT
