+++ src/waveform/widgets/deprecated/glwaveformwidgetabstract.cpp
@@ -1,4 +1,4 @@
-#include "waveform/widgets/glwaveformwidgetabstract.h"
+#include "waveform/widgets/deprecated/glwaveformwidgetabstract.h"
 
 #include <QApplication>
 #include <QWheelEvent>
